requirements.txt
numpy==1.26.4
pandas==2.2.2
sentence-transformers==2.5.1
openai==1.23.6
python-dotenv==1.0.1
transformers==4.40.2
llm2vec==0.1.6
huggingface-hub==0.21.4
matplotlib==3.8.4
seaborn==0.13.2
tqdm==4.66.2
scikit-learn==1.4.2
